Abstracted 2d chart generate methods
from ThesisScale2DQueryReportPage into anstract base class
